Moved 'Dashboard.tsx' to Phases folder and deleted the UploadResumesToJD folder
